requirements.txt
flask==3.0.2
python-dotenv==1.0.1 
openai==1.54.4
flask-cors==4.0.0
pymongo==4.10.1